separate test routines and test data for FE types
